Update mozjs to 87

https://github.com/servo/mozjs/pull/275

+++ src/typedarray.rs
@@ -111,14 +111,14 @@ impl<T: TypedArrayElement, S: JSObjectStorage> ToJSValConvertible for TypedArray
 }
 
 pub enum CreateWith<'a, T: 'a> {
-    Length(u32),
+    Length(usize),
     Slice(&'a [T]),
 }
 
 /// A typed array wrapper.
 pub struct TypedArray<T: TypedArrayElement, S: JSObjectStorage> {
     object: S,
-    computed: Cell<Option<(*mut T::Element, u32)>>,
+    computed: Cell<Option<(*mut T::Element, usize)>>,
 }
 
 unsafe impl<T> CustomTrace for TypedArray<T, *mut JSObject> where T: TypedArrayElement {
@@ -148,7 +148,7 @@ impl<T: TypedArrayElement, S: JSObjectStorage> TypedArray<T, S> {
         }
     }
 
-    fn data(&self) -> (*mut T::Element, u32) {
+    fn data(&self) -> (*mut T::Element, usize) {
         if let Some(data) = self.computed.get() {
             return data;
         }
@@ -224,7 +224,7 @@ impl<T: TypedArrayElementCreator + TypedArrayElement, S: JSObjectStorage> TypedA
                          -> Result<(), ()> {
         let length = match with {
             CreateWith::Length(len) => len,
-            CreateWith::Slice(slice) => slice.len() as u32,
+            CreateWith::Slice(slice) => slice.len(),
         };
 
         result.set(T::create_new(cx, length));
@@ -259,13 +259,13 @@ pub trait TypedArrayElement {
     /// Unwrap a typed array JS reflector for this element type.
     unsafe fn unwrap_array(obj: *mut JSObject) -> *mut JSObject;
     /// Retrieve the length and data of a typed array's buffer for this element type.
-    unsafe fn length_and_data(obj: *mut JSObject) -> (*mut Self::Element, u32);
+    unsafe fn length_and_data(obj: *mut JSObject) -> (*mut Self::Element, usize);
 }
 
 /// Internal trait for creating new typed arrays.
 pub trait TypedArrayElementCreator: TypedArrayElement {
     /// Create a new typed array.
-    unsafe fn create_new(cx: *mut JSContext, length: u32) -> *mut JSObject;
+    unsafe fn create_new(cx: *mut JSContext, length: usize) -> *mut JSObject;
     /// Get the data.
     unsafe fn get_data(obj: *mut JSObject) -> *mut Self::Element;
 }
@@ -284,7 +284,7 @@ macro_rules! typed_array_element {
                 $unwrap(obj)
             }
 
-            unsafe fn length_and_data(obj: *mut JSObject) -> (*mut Self::Element, u32) {
+            unsafe fn length_and_data(obj: *mut JSObject) -> (*mut Self::Element, usize) {
                 let mut len = 0;
                 let mut shared = false;
                 let mut data = ptr::null_mut();
@@ -304,7 +304,7 @@ macro_rules! typed_array_element {
         typed_array_element!($t, $element, $unwrap, $length_and_data);
 
         impl TypedArrayElementCreator for $t {
-            unsafe fn create_new(cx: *mut JSContext, length: u32) -> *mut JSObject {
+            unsafe fn create_new(cx: *mut JSContext, length: usize) -> *mut JSObject {
                 $create_new(cx, length)
             }
